refactor: review-comment: Move file writer to core
