+++ tests/test_cli.py
@@ -11,12 +11,12 @@ from _pytest.logging import LogCaptureFixture
 from click.testing import CliRunner, Result
 
 from opossum_lib.cli import generate
+from opossum_lib.core.services.opossum_file_writer import OpossumFileWriter
 from opossum_lib.shared.constants import (
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
 )
 from opossum_lib.shared.entities.opossum_information import OpossumPackage
-from opossum_lib.shared.services.opossum_file_writer import OpossumFileWriter
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
 
 test_data_path = Path(__file__).resolve().parent / "data"
