+++ tests/test_opossum/test_file_generation.py
@@ -5,12 +5,12 @@
 from pathlib import Path
 from zipfile import ZipFile
 
+from opossum_lib.core.services.opossum_file_writer import OpossumFileWriter
 from opossum_lib.shared.constants import (
     INPUT_JSON_NAME,
     OUTPUT_JSON_NAME,
 )
 from opossum_lib.shared.entities.opossum_file_content import OpossumFileContent
-from opossum_lib.shared.services.opossum_file_writer import OpossumFileWriter
 from tests.test_setup.opossum_file_faker_setup import OpossumFileFaker
